backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
pandas
pymysql
GitPython
openai
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
numpy
streamlit
bcrypt
streamlit-lottie
requests
